requirements.txt
fastapi
pydantic-settings
pandas
numpy
yfinance
scipy
finnhub-python
matplotlib
loguru
uvicorn
gunicorn
apscheduler
emoji
